+++ mrz-3-edition/chapters/6/exercises/6/exercise6_6.m
@@ -1,7 +1,7 @@
-function exercise3_6 ()
-% EXERCISE3_6 Solves exercse 3.6 of book.
+function exercise6_6 ()
+% EXERCISE6_6 Solves exercse 6.6 of book.
 %
-% exercise3_6 ()
+% exercise6_6 ()
 %
 % Use Newton method to solve f(x) = 0. Then use a different pt-fixed method
 % and plot the comparison..